Check if elasticsearch is already installed

+++ src/Commands/ElasticCommand.php
@@ -16,13 +16,14 @@ class ElasticCommand extends RunCommand
         return true;
     }
 
-    protected function addArguments()
-    {
-
-    }
+    protected function addArguments() {}
 
     protected function addOptions()
     {
+        $this->makeOption('install')
+            ->setDescription('Install elasticsearch in vagrant.')
+            ->boolean();
+
         $this->makeOption('up')
             ->setDescription('Boot the elasticsearch server.')
             ->boolean();
@@ -50,35 +51,52 @@ class ElasticCommand extends RunCommand
     public function execute(InputInterface $input, OutputInterface $output)
     {
         $cdToHome =  'cd Development/soapbox/soapbox-v4/ && ';
+        $isInstalled = !$this->runMyCommand('cd elasticsearch*');
 
-        if ($input->getOption('up')) {
-            $output->writeln('<info>Booting up elasticsearch...</info>');
-            $this->runMyCommand('nohup ~/elasticsearch-*/bin/elasticsearch & sleep 1');
+        if ($isInstalled) {
+            if ($input->getOption('install')) {
+                $output->writeln('<info>Elastic search is already installed!</info>');
+            }
+
+            if ($input->getOption('up')) {
+                $output->writeln('<info>Booting up elasticsearch...</info>');
+                $this->runMyCommand('nohup ~/elasticsearch-*/bin/elasticsearch & sleep 1');
+                $output->writeln('<info>Done!</info>');
+            }
+
+            if ($input->getOption('migrate')) {
+                $output->writeln('<info>Indexing documents into elasticsearch...</info>');
+                $this->runMyCommand($cdToHome.'php artisan elasticsearch:daily --reindex=true');
+                $output->writeln('<info>Done!</info>');
+            }
+
+            if ($input->getOption('refresh')) {
+                $output->writeln('<info>Deleting elasticsearch indexes...</info>');
+                $this->runMyCommand('curl -XDELETE localhost:9200/*');
+                $output->writeln('<info>Reindexing elasticsearch indexes...</info>');
+                $this->runMyCommand($cdToHome.'
+                    php artisan index:audits --add=true &&
+                    php artisan elasticsearch:daily --reindex=true &&
+                    php artisan elasticsearch:audits --mapping=true &&
+                    php artisan elasticsearch:audits --reindex=true &&
+                    php artisan index:audits --drop=true
+                ');
+                $output->writeln('<info>Done!</info>');
+            }
+
+            if ($input->getOption('halt')) {
+                $output->writeln('<info>Halting elasticsearch server...</info>');
+                $this->runMyCommand('pgrep -f elasticsearch | xargs kill -9');
+                $output->writeln('<info>Done!</info>');
+            }
+        } else {
+            if ($input->getOption('install')) {
+                $output->writeln('<info>Installing elasticsearch...</info>');
+                $output->writeln('<info>Done!</info>');
+            } else {
+                $output->writeln('<info>Please install elasticsearch first. `raven elasticsearch --install`</info>');
+            }
         }
 
-        if ($input->getOption('migrate')) {
-            $output->writeln('<info>Indexing documents into elasticsearch...</info>');
-            $this->runMyCommand($cdToHome.'php artisan elasticsearch:daily --reindex=true');
-        }
-
-        if ($input->getOption('refresh')) {
-            $output->writeln('<info>Deleting elasticsearch indexes...</info>');
-            $this->runMyCommand('curl -XDELETE localhost:9200/*');
-            $output->writeln('<info>Reindexing elasticsearch indexes...</info>');
-            $this->runMyCommand($cdToHome.'
-                php artisan index:audits --add=true &&
-                php artisan elasticsearch:daily --reindex=true &&
-                php artisan elasticsearch:audits --mapping=true &&
-                php artisan elasticsearch:audits --reindex=true &&
-                php artisan index:audits --drop=true
-            ');
-        }
-
-        if ($input->getOption('halt')) {
-            $output->writeln('<info>Halting elasticsearch server...</info>');
-            $this->runMyCommand('pgrep -f elasticsearch | xargs kill -9');
-        }
-
-        $output->writeln('<info>Done!</info>');
     }
 }
